+++ storage/src/database/mod.rs
@@ -71,10 +71,6 @@ impl<B: Backend, Sch: Schema> Storage<B, Sch> {
     pub fn backend_impl(&self) -> &B::Impl {
         &self.backend
     }
-
-    pub fn backend_impl_mut(&mut self) -> &mut B::Impl {
-        &mut self.backend
-    }
 }
 
 impl<B: Backend, Sch: Schema> Storage<B, Sch> {